vault/src/requirements.txt
Flask==0.12
PyJWT==1.4.2
pymongo==3.4.0
PyVCF==0.6.8
oic==0.9.4.0
celery==4.0.2
Pillow==4.0.0
connexion==1.1.4
PyYAML==3.12
git+https://github.com/clearnet-io/backendlibs.git
